Move 'clear session' to advanced actions
