+++ app/res/xml/advanced_actions.xml
@@ -8,4 +8,5 @@
     <Preference android:key="connection-test"/>
     <Preference android:key="recovery-mode"/>
     <Preference android:key="clear-user-data"/>
+    <Preference android:key="clear-saved-session"/>
 </PreferenceScreen>
